+++ search-films/src/components/Button/index.jsx
@@ -1,4 +1,4 @@
-import styles from './styles/index.module.css';
+import styles from './index.module.css';
 
 function Button({text}) {
 	return (