+++ search-films/src/components/MovieCard/index.jsx
@@ -1,6 +1,6 @@
 
 import { default as CssUtils } from '../shared/CssUtils';
-import styles from './styles/index.module.css';
+import styles from './index.module.css';
 
 /**
  * @callback requestCallback